Move config and package.json to agree with other repos
