+++ Dockerfile
@@ -1,0 +1,5 @@
+FROM node:16
+
+COPY . /apps/amaretti
+WORKDIR /apps/amaretti/api
+RUN npm install
