+++ api/controllers/index.js
@@ -7,7 +7,7 @@ const request = require('request');
 //const redis = require('redis');
 
 //mine
-const config = require('../../config');
+const config = require('../config');
 const logger = winston.createLogger(config.logger.winston);
 const db = require('../models');
 const common = require('../common');
