+++ bin/reset_metric.js
@@ -2,7 +2,7 @@
 const winston = require('winston');
 const async = require('async');
 
-const config = require('../config');
+const config = require('../api/config');
 const logger = winston.createLogger(config.logger.winston);
 const db = require('../api/models');
 const common = require('../api/common');
